commit aa

+++ src/main/java/com/mmall/controller/portal/CartController.java
@@ -6,12 +6,18 @@ import com.mmall.common.ServerResponse;
 import com.mmall.pojo.Cart;
 import com.mmall.pojo.User;
 import com.mmall.service.ICartService;
+import com.mmall.util.CookieUtil;
+import com.mmall.util.JsonUtil;
+import com.mmall.util.RedisPoolUtil;
 import com.mmall.vo.CartVo;
+import org.apache.commons.lang3.StringUtils;
+import org.codehaus.jackson.type.TypeReference;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
 /**
@@ -26,20 +32,30 @@ public class CartController {
 
     @RequestMapping("list.do")
     @ResponseBody
-    public ServerResponse<CartVo> list(HttpSession session){
-        User user =(User)session.getAttribute(Const.CURRENT_USER);
+    public ServerResponse<CartVo> list(HttpServletRequest httpServletRequest){
+        String cookieValue= CookieUtil.readLoginToken(httpServletRequest);
+        if(StringUtils.isEmpty(cookieValue)){
+            return ServerResponse.createByErrorMessage("查询不到用户");
+        }
+        String userJson= RedisPoolUtil.get(cookieValue);
+        User user= JsonUtil.String2Obj(userJson, new TypeReference<User>() {});
         if(user==null){
-            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),ResponseCode.NEED_LOGIN.getDesc());
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),"未登录，需要强制登录");
         }
         return iCartService.list(user.getId());
     }
 
     @RequestMapping("add.do")
     @ResponseBody
-    public ServerResponse<CartVo> add(HttpSession session, Integer count, Integer productId){
-        User user =(User)session.getAttribute(Const.CURRENT_USER);
+    public ServerResponse<CartVo> add(HttpServletRequest httpServletRequest, Integer count, Integer productId){
+        String cookieValue=CookieUtil.readLoginToken(httpServletRequest);
+        if(StringUtils.isEmpty(cookieValue)){
+            return ServerResponse.createByErrorMessage("查询不到用户");
+        }
+        String userJson=RedisPoolUtil.get(cookieValue);
+        User user=JsonUtil.String2Obj(userJson, new TypeReference<User>() {});
         if(user==null){
-            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),ResponseCode.NEED_LOGIN.getDesc());
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),"未登录，需要强制登录");
         }
         //购物车核心逻辑；
         return  iCartService.add(user.getId(),productId,count);
@@ -47,70 +63,105 @@ public class CartController {
 
     @RequestMapping("update.do")
     @ResponseBody
-    public ServerResponse<CartVo> update(HttpSession session,Integer count,Integer productId){
-        User user =(User)session.getAttribute(Const.CURRENT_USER);
+    public ServerResponse<CartVo> update(HttpServletRequest httpServletRequest,Integer count,Integer productId){
+        String cookieValue=CookieUtil.readLoginToken(httpServletRequest);
+        if(StringUtils.isEmpty(cookieValue)){
+            return ServerResponse.createByErrorMessage("查询不到用户");
+        }
+        String userJson=RedisPoolUtil.get(cookieValue);
+        User user=JsonUtil.String2Obj(userJson, new TypeReference<User>() {});
         if(user==null){
-            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),ResponseCode.NEED_LOGIN.getDesc());
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),"未登录，需要强制登录");
         }
         return iCartService.update(user.getId(),productId,count);
     }
 
     @RequestMapping("delete.do")
     @ResponseBody
-    public ServerResponse<CartVo> deleteProduct(HttpSession session,String productIds){
-        User user =(User)session.getAttribute(Const.CURRENT_USER);
+    public ServerResponse<CartVo> deleteProduct(HttpServletRequest httpServletRequest,String productIds){
+        String cookieValue=CookieUtil.readLoginToken(httpServletRequest);
+        if(StringUtils.isEmpty(cookieValue)){
+            return ServerResponse.createByErrorMessage("查询不到用户");
+        }
+        String userJson=RedisPoolUtil.get(cookieValue);
+        User user=JsonUtil.String2Obj(userJson, new TypeReference<User>() {});
         if(user==null){
-            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),ResponseCode.NEED_LOGIN.getDesc());
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),"未登录，需要强制登录");
         }
         return iCartService.deleteProduct(user.getId(),productIds);
     }
 
     @RequestMapping("select_all.do")
     @ResponseBody
-    public  ServerResponse<CartVo> selectAll(HttpSession session){
-        User user =(User)session.getAttribute(Const.CURRENT_USER);
+    public  ServerResponse<CartVo> selectAll(HttpServletRequest httpServletRequest){
+        String cookieValue=CookieUtil.readLoginToken(httpServletRequest);
+        if(StringUtils.isEmpty(cookieValue)){
+            return ServerResponse.createByErrorMessage("查询不到用户");
+        }
+        String userJson=RedisPoolUtil.get(cookieValue);
+        User user=JsonUtil.String2Obj(userJson, new TypeReference<User>() {});
         if(user==null){
-            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),ResponseCode.NEED_LOGIN.getDesc());
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),"未登录，需要强制登录");
         }
         return iCartService.selectOrUnSelect(user.getId(),null,Const.Cart.CHECKED);
     }
 
     @RequestMapping("un_select_all.do")
     @ResponseBody
-    public  ServerResponse<CartVo> unSelectAll(HttpSession session){
-        User user =(User)session.getAttribute(Const.CURRENT_USER);
+    public  ServerResponse<CartVo> unSelectAll(HttpServletRequest httpServletRequest){
+        String cookieValue=CookieUtil.readLoginToken(httpServletRequest);
+        if(StringUtils.isEmpty(cookieValue)){
+            return ServerResponse.createByErrorMessage("查询不到用户");
+        }
+        String userJson=RedisPoolUtil.get(cookieValue);
+        User user=JsonUtil.String2Obj(userJson, new TypeReference<User>() {});
         if(user==null){
-            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),ResponseCode.NEED_LOGIN.getDesc());
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),"未登录，需要强制登录");
         }
         return iCartService.selectOrUnSelect(user.getId(),null,Const.Cart.UN_CHECK);
     }
 
     @RequestMapping("un_select.do")
     @ResponseBody
-    public  ServerResponse<CartVo> unSelect(HttpSession session,Integer productId){
-        User user =(User)session.getAttribute(Const.CURRENT_USER);
+    public  ServerResponse<CartVo> unSelect(HttpServletRequest httpServletRequest,Integer productId){
+        String cookieValue=CookieUtil.readLoginToken(httpServletRequest);
+        if(StringUtils.isEmpty(cookieValue)){
+            return ServerResponse.createByErrorMessage("查询不到用户");
+        }
+        String userJson=RedisPoolUtil.get(cookieValue);
+        User user=JsonUtil.String2Obj(userJson, new TypeReference<User>() {});
         if(user==null){
-            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),ResponseCode.NEED_LOGIN.getDesc());
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),"未登录，需要强制登录");
         }
         return iCartService.selectOrUnSelect(user.getId(),productId,Const.Cart.UN_CHECK);
     }
 
     @RequestMapping("select.do")
     @ResponseBody
-    public  ServerResponse<CartVo> Select(HttpSession session,Integer productId){
-        User user =(User)session.getAttribute(Const.CURRENT_USER);
+    public  ServerResponse<CartVo> Select(HttpServletRequest httpServletRequest,Integer productId){
+        String cookieValue=CookieUtil.readLoginToken(httpServletRequest);
+        if(StringUtils.isEmpty(cookieValue)){
+            return ServerResponse.createByErrorMessage("查询不到用户");
+        }
+        String userJson=RedisPoolUtil.get(cookieValue);
+        User user=JsonUtil.String2Obj(userJson, new TypeReference<User>() {});
         if(user==null){
-            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),ResponseCode.NEED_LOGIN.getDesc());
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),"未登录，需要强制登录");
         }
         return iCartService.selectOrUnSelect(user.getId(),productId,Const.Cart.CHECKED);
     }
 
     @RequestMapping("get_cart_product_count.do")
     @ResponseBody
-    public ServerResponse<Integer> getCartProductCount(HttpSession session){
-        User user =(User)session.getAttribute(Const.CURRENT_USER);
+    public ServerResponse<Integer> getCartProductCount(HttpServletRequest httpServletRequest){
+        String cookieValue=CookieUtil.readLoginToken(httpServletRequest);
+        if(StringUtils.isEmpty(cookieValue)){
+            return ServerResponse.createByErrorMessage("查询不到用户");
+        }
+        String userJson=RedisPoolUtil.get(cookieValue);
+        User user=JsonUtil.String2Obj(userJson, new TypeReference<User>() {});
         if(user==null){
-            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),ResponseCode.NEED_LOGIN.getDesc());
+            return ServerResponse.createByErrorCodeMessage(ResponseCode.NEED_LOGIN.getCode(),"未登录，需要强制登录");
         }
 
         return iCartService.getCartProductCount(user.getId());
